Added proper socket closing functionality and added a NetworkManager class to serve as an entry point to the network code.

+++ src/managers/UserManager.java
@@ -8,7 +8,7 @@ import storage.ORM;
 
 public class UserManager {
 
-	public static final int NO_USER = -1;
+	public static final long NO_USER = -1;
 
 	private static UserManager instance = new UserManager();
